fixes - removed unsigned int

+++ hello_world/6-size.c
@@ -2,11 +2,11 @@
 
 int main(void)
 {
-printf("Size of a char: %u byte(s)\n", (unsigned int)sizeof(char));
-printf("Size of an int: %u byte(s)\n", (unsigned int)sizeof(int));
-printf("Size of a long int: %u byte(s)\n", (unsigned int)sizeof(long int));
-printf("Size of a long long int: %u byte(s)\n", (unsigned int)sizeof(long long int));
-printf("Size of a float: %u byte(s)\n", (unsigned int)sizeof(float));
+printf("Size of a char: %zu byte(s)\n", sizeof(char));
+printf("Size of an int: %zu byte(s)\n", sizeof(int));
+printf("Size of a long int: %zu byte(s)\n", sizeof(long int));
+printf("Size of a long long int: %zu byte(s)\n", sizeof(long long int));
+printf("Size of a float: %zu byte(s)\n", sizeof(float));
 
 return (0);
 }
